Fix warning in nightly about manually reimplementing `div_ceil`.
This warning only pops-up when using the marco `pool` in your app.
Just running `cargo +nightly clippy` is not enough to show this warning.
But running `cargo +nightly clippy --examples --tests` the warning
popsup because in the test and example the marco is used.

`usize::div_ceil` is used instead of `($n).div_ceil(32)` because the
compiler can not inference the type.